requirements.txt
Flask==3.0.3
neo4j==4.4.7
python-dotenv==1.0.1
requests==2.31.0
gunicorn==23.0.0